translator/requirements.txt
Django==4.0.5
SQLAlchemy==1.4.37
asgiref==3.5.2
psycopg2-binary==2.9.3
sqlparse==0.4.2
googletrans-py==4.0.0